Move the contents of PokemonRedBlueClient.py from core to the world
