+++ worlds/pokemon_rb/__init__.py
@@ -9,6 +9,7 @@ from BaseClasses import Item, MultiWorld, Tutorial, ItemClassification, Location
 from Fill import fill_restrictive, FillError, sweep_from_pool
 from worlds.AutoWorld import World, WebWorld
 from worlds.generic.Rules import add_item_rule
+from worlds.LauncherComponents import Component, SuffixIdentifier, components, Type, launch_subprocess
 from .items import item_table, item_groups
 from .locations import location_data, PokemonRBLocation
 from .regions import create_regions
@@ -22,6 +23,22 @@ from .rules import set_rules
 from .level_scaling import level_scaling
 from . import logic
 from . import poke_data
+
+
+def launch_client():
+    from .client import launch
+    launch_subprocess(launch, name="PokemonRedBlueClient")
+
+
+components.append(
+    Component(
+        "Pokémon Red and Blue Client",
+        "PokemonRedBlueClient",
+        func=launch_client,
+        component_type=Type.CLIENT,
+        file_identifier=SuffixIdentifier(".apred", ".apblue"),
+    )
+)
 
 
 class PokemonSettings(settings.Group):
